requirements.txt
category-encoders==2.2.2
imbalanced-learn==0.7.0
matplotlib==3.3.1
scikit-learn==0.23.2
seaborn==0.10.1
streamlit==0.66.0
xlrd==1.2.0
SQLAlchemy==1.3.19
PyMySQL==0.10.1
psycopg2-binary==2.8.6
python-dotenv==0.14.0